refactor: split ui.lua and comments.lua into submodules
Extract pure functions and sync logic into focused submodules to improve
code navigability and reduce module size for better AI-assisted development.

- ui.lua (1,401→~700 lines): extract ui/format.lua (18 pure format/calc
  functions) and ui/extmarks.lua (6 extmark management functions)
- comments.lua (1,200→~680 lines): extract comments/data.lua (14 pure
  data functions) and comments/sync.lua (5 GitHub API sync functions)

External interfaces unchanged — original modules act as facades that
re-export all functions from submodules. No test changes required.
